test/osmodifiertest/requirements.txt
docker == 7.1.0
libvirt-python == 10.9.0
paramiko == 3.5.0
pytest == 8.3.3
PyYAML == 6.0.2
black == 24.8.0
flake8 == 7.1.0
isort == 5.13.2
mypy == 1.13.0
types-docker == 7.1.0.20240827
types-paramiko == 3.5.0.20240928
types-PyYAML == 6.0.12.20240917